+++ src/main/java/racingcar/dao/JdbcGameDao.java
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import racingcar.dto.GameIdDTO;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
