+++ src/main/java/racingcar/dao/LocalGameDao.java
@@ -1,5 +1,7 @@
 package racingcar.dao;
 
+import racingcar.dto.GameIdDTO;
+
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
