+++ src/main/java/racingcar/dto/CarNamePositionDTO.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.dto;
 
 public class CarNamePositionDTO {
