+++ src/test/java/racingcar/dao/JdbcCarDaoTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import racingcar.dto.CarNameDTO;
+import racingcar.dto.CarNamePositionDTO;
 
 import java.util.List;
